Update 2027-01-02-block-dag.md
There are two versions of the first image, with the smaller one on top. Only differences I see are in the image files' names and sizes. This image pair is repeated after the first paragraph in the section titled "What is a DAG?" It might be useful to include the difference between transaction and block DAGs.