Update 2027-01-03-sidechains.md
Q: In the section "### What is a Sidechain and Why Would You Want One?" you refer to "chains". Does this mean that a sidechain can be pegged to a main blockchain AND one or more other sidechains? Can you put a link to a document or section thereof that shows  how sidechains touch on governance? I've seen inconsistency in the capitalization of "Super Node" and "Secure Node". Please verify which format should be used. I changed "that way" to "this way" to match the (I assume) reality that the project is closer to completion (this) than it is further from completion (that).